Section 139 - 2: HEIGHT部分の作成
Expandedでくくっていても、その中に子が入れられたら、
その子の大きさになってしまう。

But our card has now resized to the size of its child.
So remember that the way that containers work is that if they have no children, they'll try to size themselves to be as large as the parent allows,
and in this case it was stretched to the full width of the column.

But once it receives the child, it will try to resize itself to fit the size of the child.
So that's why my card has now shrunken in width.

-----
crossAxisAlignment: CrossAxisAlignment.stretch,

// フォントサイズの異なる文字を下揃え（baseline）
Row(
  mainAxisAlignment: MainAxisAlignment.center,
  crossAxisAlignment: CrossAxisAlignment.baseline,
  textBaseline: TextBaseline.alphabetic,

// 定数設定（ヘビーウェイトチャンピオン）
final kNumberTextStyle = TextStyle(
  fontSize: 50.0,
  fontWeight: FontWeight.w900,
);